Standardise content.updated event across all places that use it
Searchables, tabs, YouTube embeds and clearing tables fire a
content.updated event. They now pass in what type of object they are,
plus a copy of the object.

This means only the accordion that has been affected has to check its size.

The font.loaded event passes a 'fonts' type, resizing the width passes
a 'resize' type, and the first call to it passes an 'initial' event.

Complicated, no?